优化tabbedPage

+++ library/astator.Core/UI/Base/RadiusOutlineProvider.cs
@@ -3,8 +3,8 @@
 namespace astator.Core.UI.Base;
 public class RadiusOutlineProvider : ViewOutlineProvider
 {
-    private readonly int radius;
-    public RadiusOutlineProvider(int radius)
+    private readonly float radius;
+    public RadiusOutlineProvider(float radius)
     {
         this.radius = radius;
     }
